refactor: unified ArchiveStats from events/snapshots with h1_live fallback
Replace SeasonStats + CombatStats + EventStats with one ArchiveStats
component. Always shows event/snapshot-derived stats (works for all
seasons). Adds h1_live combat stats when available (current season).
